@@ -15,15 +15,15 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
-		<maven.compiler.target>${java.version}</maven.compiler.target>
-		<maven.compiler.source>${java.version}</maven.compiler.source>
-		<maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
-		<maven-release-plugin.version>2.5.3</maven-release-plugin.version>
+		<java.version>17</java.version>
+		<maven.compiler.target>17</maven.compiler.target>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven-gpg-plugin.version>3.2.1</maven-gpg-plugin.version>
+		<maven-release-plugin.version>3.0.1</maven-release-plugin.version>
 		<nexus-staging-maven-plugin>1.6.8</nexus-staging-maven-plugin>
 		<maven-javadoc-plugin.version>3.0.1</maven-javadoc-plugin.version>
-		<spring-boot.version>2.0.9.RELEASE</spring-boot.version>
-		<maven-source-plugin.version>3.0.1</maven-source-plugin.version>
+		<spring-boot.version>3.2.4</spring-boot.version>
+		<maven-source-plugin.version>3.3.0</maven-source-plugin.version>
 	</properties>
 
 	<dependencyManagement>
@@ -42,8 +42,16 @@
 		<dependency>
 			<groupId>org.apache.maven</groupId>
 			<artifactId>maven-plugin-api</artifactId>
-			<version>3.0</version>
+			<version>3.9.6</version>
 		</dependency>
+		<!-- https://mvnrepository.com/artifact/org.apache.maven/maven-project -->
+		<dependency>
+			<groupId>org.apache.maven</groupId>
+			<artifactId>maven-project</artifactId>
+			<version>2.2.1</version>
+		</dependency>
+
+
 		<dependency>
 			<groupId>org.apache.maven.plugin-tools</groupId>
 			<artifactId>maven-plugin-annotations</artifactId>
@@ -71,7 +79,7 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-plugin-plugin</artifactId>
-				<version>3.6.0</version>
+				<version>3.10.2</version>
 				<configuration>
 					<goalPrefix>chenile-config</goalPrefix>
 					<skipErrorNoDescriptorsFound>true</skipErrorNoDescriptorsFound>
