@@ -35,12 +35,12 @@
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<!-- version>3.2.3</version -->
-		<version>2.1.6.RELEASE</version>
+		<version>3.2.4</version>
 	</parent>
 
 	<properties>
-		<java.version>8</java.version >
-		<java.compiler.version>1.8</java.compiler.version>
+		<java.version>17</java.version >
+<!--		<java.compiler.version>1.8</java.compiler.version>-->
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<!-- tomcat.version>9.0.21</tomcat.version>
 		<zookeeper.version>3.4.13</zookeeper.version>
@@ -48,7 +48,7 @@
 	
 		<!-- java.version>21</java.version -->
 		<cucumber.version>4.3.1</cucumber.version>
-		<jacoco.version>0.8.5</jacoco.version>
+		<jacoco.version>0.8.11</jacoco.version>
 		<!-- revision>0.0.9-SNAPSHOT</revision -->
 		<chenile.version>${revision}</chenile.version>
 		<springdoc.version>1.5.9</springdoc.version>
