@@ -16,8 +16,8 @@
 	<name>Chenile State Transition Machine</name>
 
 	<properties>
-	    <maven.compiler.source>8</maven.compiler.source>
-	    <maven.compiler.target>8</maven.compiler.target>
+	    <maven.compiler.source>17</maven.compiler.source>
+	    <maven.compiler.target>17</maven.compiler.target>
 	</properties>
 
 	<dependencies>
